% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorToChar.R
\name{factorToChar}
\alias{factorToChar}
\title{Convert columns from "factor" to "character"}
\usage{
factorToChar(data)
}
\arguments{
\item{data}{Working dataset for conversion.}
}
\description{
This function converts all columns of class "factor" into "character"
}
\examples{
test <- iris #< Set working dataset
sapply(test, class)
testConv <- factorToChar(test)
sapply(testConv, class)
}
\keyword{character}
\keyword{convert,}
\keyword{factor,}
